+++ src/app/compare/Compare.tsx
@@ -5,7 +5,6 @@ import { DimItem, DimStat } from '../inventory/item-types';
 import { router } from '../../router';
 import * as _ from 'lodash';
 import { CompareService } from './compare.service';
-import { toaster } from '../ngimport-more';
 import { chainComparator, reverseComparator, compareBy } from '../comparators';
 import { createSelector } from 'reselect';
 import CompareItem from './CompareItem';
@@ -15,6 +14,7 @@ import { connect } from 'react-redux';
 import { ReviewsState, getRating, ratingsSelector } from '../item-review/reducer';
 import { RootState } from '../store/reducers';
 import Sheet from '../dim-ui/Sheet';
+import { showNotification } from '../notifications/notifications';
 
 interface StoreProps {
   ratings: ReviewsState['ratings'];
@@ -72,7 +72,7 @@ class Compare extends React.Component<Props, State> {
     });
 
     this.subscriptions.add(
-      CompareService.compareItem$.subscribe((args) => {
+      CompareService.compareItems$.subscribe((args) => {
         this.setState({ show: true });
         CompareService.dialogOpen = true;
 
@@ -214,7 +214,10 @@ class Compare extends React.Component<Props, State> {
     this.setState({ sortedHash });
   };
 
-  private add = ({ item, dupes }: { item: DimItem; dupes: boolean }) => {
+  private add = ({ items, dupes }: { items: DimItem[]; dupes: boolean }) => {
+    // use the first item and assume all others are of the same 'type'
+    const item = items[0];
+
     if (!item.comparable) {
       return;
     }
@@ -226,19 +229,23 @@ class Compare extends React.Component<Props, State> {
       comparisons[0].typeName &&
       item.typeName !== comparisons[0].typeName
     ) {
-      toaster.pop(
-        'warning',
-        item.name,
-        comparisons[0].classType && item.classType !== comparisons[0].classType
-          ? t('Compare.Error.Class', { class: comparisons[0].classTypeNameLocalized })
-          : t('Compare.Error.Archetype', { type: comparisons[0].typeName })
-      );
+      showNotification({
+        type: 'warning',
+        title: item.name,
+        body:
+          comparisons[0].classType && item.classType !== comparisons[0].classType
+            ? t('Compare.Error.Class', { class: comparisons[0].classTypeNameLocalized })
+            : t('Compare.Error.Archetype', { type: comparisons[0].typeName })
+      });
       return;
     }
 
-    if (dupes) {
-      const allItems = item.getStoresService().getAllItems();
-      const similarTypes = this.findSimilarTypes(allItems, item);
+    const allItems = item.getStoresService().getAllItems();
+    const similarTypes = this.findSimilarTypes(allItems, item);
+
+    if (items.length > 1) {
+      this.setState({ similarTypes, archetypes: [], comparisons: [...comparisons, ...items] });
+    } else if (dupes) {
       const archetypes = this.findArchetypes(similarTypes, item);
       this.setState({
         comparisons: allItems.filter((i) => i.hash === item.hash),