+++ src/app/compare/compare.service.ts
@@ -3,11 +3,11 @@ import { Subject } from 'rxjs/Subject';
 
 export const CompareService = {
   dialogOpen: false,
-  compareItem$: new Subject<{
-    item: DimItem;
+  compareItems$: new Subject<{
+    items: DimItem[];
     dupes: boolean;
   }>(),
-  addItemToCompare(item: DimItem, dupes = false) {
-    this.compareItem$.next({ item, dupes });
+  addItemsToCompare(items: DimItem[], dupes = false) {
+    this.compareItems$.next({ items, dupes });
   }
 };