+++ src/app/d2-loadout-builder/locked-armor/LockedItemContainer.tsx
@@ -1,12 +1,12 @@
 import { DestinyInventoryItemDefinition } from 'bungie-api-ts/destiny2';
 import * as React from 'react';
-import BungieImage from '../../dim-ui/BungieImage';
 import { InventoryBucket } from '../../inventory/inventory-buckets';
 import { D2Item } from '../../inventory/item-types';
 import { toggleLockedItem } from '../generated-sets/utils';
 import LoadoutBuilderItem from '../LoadoutBuilderItem';
 import { BurnItem, LockedItemType } from '../types';
 import './lockeditemcontainer.scss';
+import BungieImageAndAmmo from '../../dim-ui/BungieImageAndAmmo';
 
 /**
  * Render the locked item bucket. Could contain an item, perk, burn
@@ -52,8 +52,9 @@ export default function LockedItemContainer({
       case 'perk':
         const perkItem = lockedItem.item as DestinyInventoryItemDefinition;
         return (
-          <BungieImage
+          <BungieImageAndAmmo
             key={perkItem.index}
+            hash={perkItem.hash}
             className="empty-item"
             title={perkItem.displayProperties.name}
             src={perkItem.displayProperties.icon}