+++ src/app/inventory/dimCsvService.factory.ts
@@ -309,6 +309,7 @@ function downloadWeapons(items: DimItem[], nameMap: { [key: string]: string }) {
           switch (stat.statHash) {
             case 2715839340: // Recoil direction
               stats.recoil = stat.value;
+              break;
             case 1345609583: // Aim Assist
               stats.aa = stat.value;
               break;